:bug: fix: rename folder name
